+++ src/super/i-block/modules/dom/interface.ts
@@ -11,3 +11,12 @@ import iBlock from 'super/i-block/i-block';
 export interface ElCb<CTX extends iBlock = iBlock> {
 	(this: CTX, el: Element): any;
 }
+
+export const
+	inViewInstanceStoreSymbol: unique symbol = Symbol.for('in-view instance store');
+
+/**
+ * @see https://github.com/microsoft/TypeScript/issues/1863
+ */
+export const
+	inViewInstanceStore = <any>inViewInstanceStoreSymbol;